requirements.txt
cdsapi>=0.5
click>=8.0
dotmap>=1.3
Flask>=2.2
pandas>=1.3
PyYAML>=6.0
requests>=2.28
SQLAlchemy>=1.4, <2.0
WSGIserver>=1.3
xarray>=2022.11
sphinx_rtd_theme
